Move ClusterLoadAssignment to pkg/eds/cla
