+++ pkg/envoy/eds/types.go
@@ -1,5 +1,1 @@
 package eds
-
-const (
-	clusterLoadAssignment = "type.googleapis.com/envoy.api.v2.ClusterLoadAssignment"
-)
